감정 타임머신 리스트 삭제 버그 수정
감정 타임머신 리스트 삭제 버그 수정

+++ app/src/main/java/com/depromeet/bboxx/presentation/ui/feelhistory/UserClickEvent.kt
@@ -4,8 +4,8 @@ import com.depromeet.bboxx.presentation.model.NotificationModel
 
 interface UserClickEvent {
 
-    fun onItemClick(notifications: NotificationModel, position: Long)
+    fun onItemClick(notifications: NotificationModel, position: Int)
 
-    fun onItemDeleteClick(notifications: NotificationModel, position: Long)
+    fun onItemDeleteClick(notifications: NotificationModel, position: Int)
 
 }
